Quantities feature working and exporting JSON
