+++ src/main.ts
@@ -2,7 +2,6 @@ import * as THREE from "three";
 import * as OBC from "@thatopen/components";
 import * as OBF from "@thatopen/components-front";
 import * as BUI from "@thatopen/ui";
-import { FragmentsGroup } from "@thatopen/fragments";
 import projectInformation from "./components/Panels/ProjectInformation";
 import elementData from "./components/Panels/Selection";
 import settings from "./components/Panels/Settings";
@@ -11,9 +10,9 @@ import camera from "./components/Toolbars/Sections/Camera";
 import selection from "./components/Toolbars/Sections/Selection";
 import { AppManager } from "./bim-components";
 import { SimpleQTO } from "./bim-components/SimpleQTO copy/src/SimpleQTO";
-import { qtoTool } from "./bim-components/SimpleQTO copy/src/Template";
 
 import "./style.css";
+import QTO from "./components/Panels/QTO";
 
 // Initialize application
 (async () => {
@@ -119,6 +118,7 @@ import "./style.css";
     // Setup UI components
     const projectInformationPanel = projectInformation(components);
     const elementDataPanel = elementData(components);
+    const qtoPanel = QTO(components);
 
     const leftPanel = BUI.Component.create(() => {
       return BUI.html`
@@ -152,33 +152,21 @@ import "./style.css";
       }
     };
 
-    const qtoTable = qtoTool({ components });
-    const qtoPanel = BUI.Component.create<BUI.Panel>(() => {
-      return BUI.html`
-        <bim-panel>
-            <bim-panel-section
-             name="qto"
-             label="Quantities"
-             icon="solar:document-bold"
-             fixed
-            >
-                ${qtoTable}
-            </bim-panel-section>
-        </bim-panel>
-      `;
-    });
-
     const toolbar = BUI.Component.create(() => {
       return BUI.html`
         <bim-toolbar>
           ${load(components)}
           ${camera(world)}
           ${selection(components, world)}
-          <bim-button 
-            tooltip-title="Quantities" 
+          <bim-toolbar-section label="Quantities" icon="solar:import-bold">
+            <bim-button 
+            label="Total Quantities"
+            tooltip-title="Sum quantities" 
+            tooltip-text="Adds up the quantities of the selected elements"
             icon="mdi:summation"
             @click=${onShowQuantity}
-          ></bim-button>
+            ></bim-button>
+          </bim-toolbar-section>
         </bim-toolbar>
       `;
     });
